fix(protocol): unixOwnershipEqual: allow Ignoring OwnerName
When 2 linux systems with mismatched UserNames and !IgnoreOwnership sync,
for example device A with User: A & UID: 100, and Device B with User: B and UID: 100.
Critically device B does NOT have User: A. 
when syncing currently the receiving device (B) will lookup User: A, not find it, and instead just use the UID: 1000.
when the device (B) does its next scan it will find a global file with (User: A & UID: 100), But a local file with (User: B & UID: 100) as pilled from the Fs. This shows up as a conflict, but it can never be resolved. As every time it sybcs the same UID will be applied.
This is primarily seen when Device B is a Receive Only Folder.

My use case is using syncthing between 2 linux Devies, with one set to Send Only and the other Receive Only.

For prior art see: https://github.com/syncthing/syncthing/issues/9039#issuecomment-1870589955